# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"RepositoryDTO_unknown_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {},
					"data": {}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteModel": {
				"properties": {
					"ids": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"ids"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RepositoryDTO_any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {},
					"data": {}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TypeSort": {
				"enum": [
					0,
					1
				],
				"type": "number"
			},
...
			"UserModel": {
				"properties": {
					"spice_preference": {
						"type": "number",
						"format": "double"
					},
					"sweetness_preference": {
						"type": "number",
						"format": "double"
					},
					"saltiness_preference": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"regionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"spice_preference",
					"sweetness_preference",
					"saltiness_preference",
					"description",
					"regionId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RecipeIngredientModel": {
				"properties": {
					"ingredientId": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unit": {
						"type": "string"
					}
				},
				"required": [
					"ingredientId",
					"quantity",
					"unit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RecipeModel": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"instructions": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"categoryId": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"regionId": {
						"type": "number",
						"format": "double"
					},
					"recipeIngredient": {
						"items": {
							"$ref": "#/components/schemas/RecipeIngredientModel"
						},
						"type": "array"
					},
					"timeCook": {
						"type": "number",
						"format": "double"
					},
					"spice_level": {
						"type": "number",
						"format": "double"
					},
					"sweetness_level": {
						"type": "number",
						"format": "double"
					},
					"saltiness_level": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"description",
					"instructions",
					"imageUrl",
					"categoryId",
					"regionId",
					"recipeIngredient",
					"timeCook",
					"spice_level",
					"sweetness_level",
					"saltiness_level"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RatingRequest": {
				"properties": {
					"comment": {
						"type": "string"
					},
					"rate": {
						"type": "number",
						"format": "double"
					},
					"recipeId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"comment",
					"rate",
					"recipeId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IngredientModel": {
				"properties": {
					"name": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"imageUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FavouriteModel": {
				"properties": {
					"recipeId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"recipeId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CategoryModel": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"JWT": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": ""
	},
	"paths": {
		"/user": {
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"JWT": [
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"JWT": [
							null
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "username",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"JWT": [
							null,
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserModel"
							}
						}
					}
				}
			}
		},
...
		"/recipe": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Recipe"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "ingredientIds",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "categoryIds",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Recipe"
				],
				"security": [
					{
						"JWT": [
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RecipeModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Recipe"
				],
				"security": [
					{
						"JWT": [
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
...
		"/rating": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"description": "Lọc rating  tên,page,pagesize,với sắp xếp",
				"tags": [
					"Rating"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "recipeId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "star",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Create",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Rating"
				],
				"security": [
					{
						"JWT": [
							null,
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RatingRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"204": {
						"description": "No content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Rating"
				],
				"security": [
					{
						"JWT": [
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
		"/rating/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Rating"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Rating"
				],
				"security": [
					{
						"JWT": [
							null,
							null
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RatingRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Rating"
				],
				"security": [
					{
						"JWT": [
							null,
							null
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/rating/getRatingMe/{recipeId}": {
			"get": {
				"operationId": "GetRatingMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Rating"
				],
				"security": [
					{
						"JWT": [
							null,
							null
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "recipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/ingredient": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "trả về một danh sách món ăn",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredient"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredient"
				],
				"security": [
					{
						"JWT": [
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IngredientModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Ingredient"
				],
				"security": [
					{
						"JWT": [
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
...
		"/favourite": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"favourite"
				],
				"security": [
					{
						"JWT": [
							null,
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FavouriteModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveArray",
				"responses": {
					"200": {
						"description": "Xóa đồ ăn khỏi danh sách yêu thích",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"favourite"
				],
				"security": [
					{
						"JWT": [
							null,
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"favourite"
				],
				"security": [
					{
						"JWT": [
							null,
							null
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/category": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"description": "Lọc thể loại đồ ăn thêm tên,page,pagesize,với sắp xếp",
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Create",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [
					{
						"JWT": [
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"204": {
						"description": "No content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [
					{
						"JWT": [
							null
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}